Examples now working
